used progress property on goal to show progress on home view controller

+++ Hijack/Models/Goal.swift
@@ -15,7 +15,7 @@ struct Goal {
     let goalId: String
     let imageURL: String
     let status: String
-    let progress: Int
+    let progress: Float
     let tasks: [Task]
     let createdDate: Timestamp
     
@@ -31,7 +31,7 @@ extension Goal {
         self.goalId = dictionary["goalId"] as? String ?? "no goal id"
         self.imageURL = dictionary["imageURL"] as? String ?? "no image URL"
         self.status = dictionary["status"] as? String ?? "no status available"
-        self.progress = dictionary["progress"] as? Int ?? 0
+        self.progress = dictionary["progress"] as? Float ?? 0
         self.tasks = dictionary["tasks"] as? [Task] ?? [Task(description: "no tasks", status: false, taskId: "no taskId", createdDate: Timestamp(date: Date()))]
         self.createdDate = dictionary["createdDate"] as? Timestamp ?? Timestamp(date: Date())
     }
